+++ docs/version_history.md
@@ -15,6 +15,7 @@ Month Day, Year
 3. refactor:
    1. refactor: __init_subclass__ in base encounter for automatic class registration
    2. refactor: BaseEncounter is an ABC
+   3. refactor: type annotations for BaseEncounter methods
 
 ### Coming soon:
 * make sure that new moves to automatically created fighters are assigned fairly (without premature bonuses)